Add the new blog.
This has almost the complete redesign. The root page redirects to the
first post, and we use related posts now to get other links. There are
two subscription links. The pitch page needs a little work, at least to
get them back to the blog, and a better call to action. That said I can
post now.